moved helper functions from minimize to common
