+++ kernel_tuner/strategies/diff_evo.py
@@ -4,8 +4,8 @@ from collections import OrderedDict
 from kernel_tuner import util
 from kernel_tuner.searchspace import Searchspace
 from kernel_tuner.strategies import common
-from kernel_tuner.strategies.minimize import (_cost_func, get_bounds,
-                                              scale_from_params)
+from kernel_tuner.strategies.common import (_cost_func, get_bounds,
+                                            scale_from_params)
 from scipy.optimize import differential_evolution
 
 supported_methods = ["best1bin", "best1exp", "rand1exp", "randtobest1exp",
