+++ kernel_tuner/strategies/random_sample.py
@@ -5,7 +5,7 @@ import numpy as np
 from kernel_tuner import util
 from kernel_tuner.searchspace import Searchspace
 from kernel_tuner.strategies import common
-from kernel_tuner.strategies.minimize import _cost_func
+from kernel_tuner.strategies.common import _cost_func
 
 _options = OrderedDict(fraction=("Fraction of the search space to cover value in [0, 1]", 0.1))
