+++ kernel_tuner/strategies/simulated_annealing.py
@@ -7,7 +7,7 @@ import numpy as np
 from kernel_tuner import util
 from kernel_tuner.searchspace import Searchspace
 from kernel_tuner.strategies import common
-from kernel_tuner.strategies.minimize import _cost_func
+from kernel_tuner.strategies.common import _cost_func
 
 _options = OrderedDict(T=("Starting temperature", 1.0),
                        T_min=("End temperature", 0.001),
